chore(metrics): Initial improvement to text_predict implementation.
