+++ archai/nlp/metrics/text_predict/wrappers/vocab_wrapper.py
@@ -6,8 +6,15 @@
 
 import functools
 import logging
+import json
+import os
+import re
 from typing import List, Optional, Set, Tuple
 
+from tokenizers import ByteLevelBPETokenizer
+from tokenizers import Tokenizer
+
+from transformers import AutoTokenizer
 import numpy as np
 
 from archai.common.lru_cache import LRUCache
@@ -16,6 +23,12 @@ from archai.nlp.metrics.text_predict.text_predict_utils import RE_SPLIT, WORD_TO
 
 # Token-related constants
 TOKENIZER_FILTER_TOKEN_IDS_CACHE_SIZE = 65536
+TOKENIZER_BOS = '_BOS_'
+TOKENIZER_MAPPING = {'<|unk|>': '_OOV_', '<|endoftext|>': TOKENIZER_BOS}
+RE_WHITESPACE = re.compile("[\xa0 \t\u2002-\u2006\u200B]+", re.MULTILINE | re.DOTALL)
+# This RE_NEW_LINE should (hopefully) make new line marker platform-independent
+RE_NEW_LINE = re.compile("\s*[\r\n]+\s*", re.MULTILINE | re.DOTALL) # pylint: disable=anomalous-backslash-in-string
+
 
 
 class VocabWrapper:
@@ -29,15 +42,78 @@ class VocabWrapper:
     # Text to insert at the beginning of each sequence
     BOS_TEXT = ''
 
-    def __init__(self, vocab: VocabBase) -> None:
-        self.vocab = vocab
+    def __init__(self, vocab_path: str) -> None:
+        # merges_file, vocab_json_file, _ = self._convert_to_separate_files(vocab_path)
 
-        self.idx2token = self.vocab.ids_to_tokens(list(range(len(self.vocab))))
+        self.tokenizer = AutoTokenizer.from_pretrained(vocab_path)
+
+
+        print(dir(self.tokenizer))
+        # print(self.tokenizer.vocab)
+
+        self.bos_token = self.tokenizer.bos_token_id
         self.filter_token_ids_cache = LRUCache(TOKENIZER_FILTER_TOKEN_IDS_CACHE_SIZE)
         
-        self.WORD_TOKEN_SEPARATOR_IDX = set([idx for idx in range(len(self)) if self[idx][0] in WORD_TOKEN_SEPARATOR_SET])
+        self.WORD_TOKEN_SEPARATOR_IDX = set([idx for idx in range(len(self)) if self[idx] in WORD_TOKEN_SEPARATOR_SET])
 
-        logging.debug(f'WORD_TOKEN_SEPARATOR_IDX size: {len(self.WORD_TOKEN_SEPARATOR_IDX)}')
+        # logging.debug(f'WORD_TOKEN_SEPARATOR_IDX size: {len(self.WORD_TOKEN_SEPARATOR_IDX)}')
+
+    def _convert_to_separate_files(self, vocab_path: str) -> Tuple[str, str]:
+        vocab_dir = os.path.dirname(vocab_path)
+        merges_path = f'{vocab_dir}/merges.txt'
+        vocab_json_path = f'{vocab_dir}/vocab.json'
+        vocab_txt_path = f'{vocab_dir}/vocab.txt'
+
+        with open(vocab_path, 'r', encoding='utf-8') as f:
+            vocab = json.load(f)
+        
+        # Saves the vocab.json
+        with open(vocab_json_path, 'w', encoding='utf-8') as f:
+	        json.dump(vocab['model']['vocab'], f)
+
+        # Saves the merges.json
+        with open(merges_path, 'w', encoding='utf-8') as f:
+            for i, merge in enumerate(vocab['model']['merges']):
+                f.write(merge)
+                if i != len(vocab['model']['merges']) - 1:
+                    f.write('\n')
+
+        # Saves the vocab.txt
+        with open(vocab_txt_path, 'w', encoding='utf-8') as f:
+            vocab_list = [0] * len(vocab['model']['vocab'])
+
+            for token, idx in vocab['model']['vocab'].items():
+                if token in TOKENIZER_MAPPING:
+                    token = TOKENIZER_MAPPING[token]
+                vocab_list[idx] = token
+
+            for i, token in enumerate(vocab_list):
+                f.write('%s' % (token))
+                
+                if i != len(vocab_list) - 1:
+                    f.write('\n')
+
+        return merges_path, vocab_json_path, vocab_txt_path
+
+    def clean(self, text: str, add_bos_text=True) -> str:
+        """Clean the text before tokenizing."""
+        # One could code it up similar to this in C++:
+        # https://stackoverflow.com/questions/40690460/python-removing-extra-special-unicode-characters
+        # logging.debug(f"tokenizer.clean.before: '{text}'")
+        # UnicodeHyphen,"[\u2014\u2015]",local.replace,,"-"
+        text = re.sub(r"[\u2010\u2011\u2012]", "-", text)
+        text = re.sub(r"\s*[\u2013\u2014\u2015]\s*", " - ", text)
+        # UnicodeQuote,"[\u2018\u2019\xb4]",local.replace,,"'"
+        text = re.sub(r"[\u2018\u2019\u201a\u201b\xb4]", "'", text)
+        # UnicodeDoubleQuote,"[\u201c\u201d]",local.replace,,""""
+        text = re.sub(r"[\u201c\u201d\u201e\u201f]", '"', text)
+
+        text = RE_WHITESPACE.sub(" ", text)
+        text = RE_NEW_LINE.sub("\n ", text)
+        if add_bos_text:
+            text = self.BOS_TEXT + text.lstrip()
+        logging.debug(f"tokenizer.clean.after {add_bos_text}: '{text}'")
+        return text
 
     @functools.lru_cache(maxsize=32768)
     def filter_token_mask_ids(self, filter_prefix: str) -> np.ndarray:
@@ -53,10 +129,11 @@ class VocabWrapper:
 
     @functools.lru_cache(maxsize=128)
     def encode(self, text: str) -> List[int]:
-        return self.vocab.encode_text(text)
+        print(f'encode: {text} -> {self.tokenizer.encode(text)}')
+        return self.tokenizer.encode(text)
 
     def decode(self, input_ids: List[int]) -> str:
-        return self.vocab.decode_text(input_ids)
+        return self.tokenizer.decode(input_ids)
 
     @functools.lru_cache(maxsize=32768)
     def filter_token_ids(self, filter_prefix: str) -> Tuple[int, ...]:
@@ -66,6 +143,8 @@ class VocabWrapper:
 
     def find_context_prefix(self, text: str) -> Tuple[str, str]:
         m = RE_SPLIT.match(text)
+
+        # print(text, m)
 
         if m is None:
             context = ''
@@ -78,13 +157,13 @@ class VocabWrapper:
         return (context, prefix)
 
     def __iter__(self) -> int:
-        yield from self.idx2token
+        yield from self.tokenizer.vocab
 
     def __len__(self) -> int:
-        return len(self.idx2token)
+        return len(self.tokenizer.vocab)
 
     def __getitem__(self, idx: int) -> str:
-        return self.idx2token[idx]
+        return self.tokenizer.decode(idx)
 
     def _filter_token_ids(self,
                           filter_prefix: str,
@@ -113,7 +192,7 @@ class VocabWrapper:
         return result
 
     @property
-    def UPPER_TOKENS(self) -> Set:
+    def upper_tokens(self) -> Set:
         if not hasattr(self, '_upper_tokens'):
             self._upper_tokens = {idx for idx in range(len(self)) if any([c.isupper() and c not in WORD_TOKEN_SEPARATOR_SET for c in self[idx]])} # pylint: disable=attribute-defined-outside-init
 
@@ -125,9 +204,3 @@ class VocabWrapper:
         result = tuple((idx,) for idx in result)
 
         return result
-
-    def clean(self, text: str, add_bos_text: Optional[bool] = True) -> str:
-        if add_bos_text:
-            text = self.BOS_TEXT + text
-
-        return text
